Adding Schema Validations and minor edits
