+++ Covinator/.vs/Covinator/config/applicationhost.config
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0" encoding="UTF-8"?>
 <!--
 
     IIS configuration sections.
@@ -157,7 +157,7 @@
       </site>
       <site name="Covinator" id="2">
         <application path="/" applicationPool="Covinator AppPool">
-          <virtualDirectory path="/" physicalPath="D:\IS7024\Covinator\Covinator" />
+          <virtualDirectory path="/" physicalPath="C:\Users\Moiz\OneDrive\Documents\GitHub\IS7024\IS7024\Covinator\Covinator" />
         </application>
         <bindings>
           <binding protocol="http" bindingInformation="*:54264:localhost" />
