+++ Covinator/Covinator/Pages/PfizerVaccination.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model Covinator.Pages.PfzierVaccination
+@model Covinator.Pages.PfizerVaccination
 @{
     ViewData["Title"] = "Home page";
     var pfizerVaccineDistributionAllocationss = (QuickType.PfizerVaccineDistributionAllocations[])ViewData["PfizerVaccineDistributionAllocations"];
