+++ Covinator/Covinator/bin/Debug/netcoreapp3.1/Covinator.runtimeconfig.json
@@ -6,7 +6,8 @@
       "version": "3.1.0"
     },
     "configProperties": {
-      "System.GC.Server": true
+      "System.GC.Server": true,
+      "System.Runtime.Serialization.EnableUnsafeBinaryFormatterSerialization": false
     }
   }
 }
