+++ Covinator/Covinator/casesDeathsSchema.json
@@ -1,0 +1,120 @@
+﻿{
+  "$schema": "http://json-schema.org/draft-06/schema#",
+  "type": "array",
+  "title": "Covid 19 Active cases and deaths ",
+  "descrption": "The data related to the number of active cases of Covid 19 and the deaths caused by it in the USA",
+  "items": {
+    "$ref": "#/definitions/CasesDeaths"
+  },
+  "definitions": {
+    "CasesDeaths": {
+      "type": "object",
+      "additionalProperties": false,
+      "properties": {
+        "submission_date": {
+          "format": "date-time",
+          "type": "string"
+        },
+        "state": {
+          "$ref": "#/definitions/State"
+        },
+        "tot_cases": {
+          "type": "string",
+          "format": "integer"
+        },
+        "conf_cases": {
+          "type": "string"
+        },
+        "prob_cases": {
+          "type": "string",
+          "format": "integer"
+        },
+        "new_case": {
+          "type": "string"
+        },
+        "pnew_case": {
+          "type": "string",
+          "format": "integer"
+        },
+        "tot_death": {
+          "type": "string",
+          "format": "integer"
+        },
+        "conf_death": {
+          "type": "string"
+        },
+        "prob_death": {
+          "type": "string",
+          "format": "integer"
+        },
+        "new_death": {
+          "type": "string"
+        },
+        "pnew_death": {
+          "type": "string",
+          "format": "integer"
+        },
+        "created_at": {
+          "type": "string",
+          "format": "date-time"
+        },
+        "consent_cases": {
+          "$ref": "#/definitions/Consent"
+        },
+        "consent_deaths": {
+          "$ref": "#/definitions/Consent"
+        }
+      },
+      "required": [
+        "created_at",
+        "new_case",
+        "new_death",
+        "state",
+        "submission_date",
+        "tot_cases",
+        "tot_death"
+      ],
+      "title": "CasesDeaths"
+    },
+    "Consent": {
+      "type": "string",
+      "enum": [
+        "Agree",
+        "N/A",
+        "Not agree"
+      ],
+      "title": "Consent"
+    },
+    "State": {
+      "type": "string",
+      "enum": [
+        "OK",
+        "SD",
+        "UT",
+        "GA",
+        "WV",
+        "PA",
+        "TN",
+        "AS",
+        "PW",
+        "HI",
+        "MP",
+        "IA",
+        "CO",
+        "KS",
+        "PR",
+        "RMI",
+        "NY",
+        "NM",
+        "AR",
+        "NYC",
+        "OH",
+        "AK",
+        "MA",
+        "TX",
+        "FL"
+      ],
+      "title": "State"
+    }
+  }
+}
