+++ Covinator/Covinator/pfizerSchema.json
@@ -1,0 +1,47 @@
+﻿{
+  "$schema": "http://json-schema.org/draft-06/schema#",
+  "title": "Pfizer allocation based on states schema ",
+  "descrption": "The allocation of pfizer vaccines based on states in the United States",
+  "type": "array",
+  "items": {
+    "$ref": "#/definitions/PfizerVaccineDistributionAllocations"
+  },
+  "definitions": {
+    "PfizerVaccineDistributionAllocations": {
+      "type": "object",
+      "additionalProperties": false,
+      "properties": {
+        "jurisdiction": {
+          "type": "string",
+          "description": "The State for which the allocation is given"
+        },
+        "week_of_allocations": {
+          "type": "string",
+          "format": "date-time",
+          "description": "The Date for the allocation"
+        },
+        "_1st_dose_allocations": {
+          "type": "string",
+          "format": "integer",
+          "description": "The number of vaccines allocated for first Dose",
+          "minimum": 0
+
+        },
+        "_2nd_dose_allocations": {
+          "type": "string",
+          "format": "integer",
+          "description": "The number of vaccines allocated for second Dose",
+          "minimum": 0
+        }
+      },
+      "required": [
+        "_1st_dose_allocations",
+        "_2nd_dose_allocations",
+        "jurisdiction",
+        "week_of_allocations"
+      ],
+      "title": "PfizerVaccineDistributionAllocations"
+    }
+  }
+}
+
